integrate methods for buying/selling in minitrader

+++ capital_service_bodies.go
@@ -10,6 +10,6 @@ type CreateWorkingOrderBody struct {
 	Epic      string    `json:"epic"`
 	Direction Signal    `json:"direction"`
 	Type      OrderType `json:"type"`
-	Size      float32   `json:"size"`
-	Level     float32   `json:"level"`
+	Size      float64   `json:"size"`
+	Level     float64   `json:"level"`
 }
